Add more errors

+++ backend/commands/admin/purge.js
@@ -1,6 +1,8 @@
-// Errors
-import IllegalArgumentError from '../../errors/IllegalArgumentError.js';
 import {success} from '../../utils/messages.js';
+
+// Errors
+import IntegerConversionError from '../../errors/IntegerConversionError.js';
+import IntegerRangeError from '../../errors/IntegerRangeError.js';
 
 
 export default {
@@ -14,10 +16,10 @@ export default {
     async execute(message, parsed) {
         let count = Number(parsed.count);
 
-        if (!count)
-            throw new IllegalArgumentError(this.name, '`Count` must be a valid integer');
+        if (isNaN(count) || count % 1 !== 0)
+            throw new IntegerConversionError(this.name, 'Count');
         if (count < 1 || count > 100)
-            throw new IllegalArgumentError(this.name, '`Count` must be between 1 and 100');
+            throw new IntegerRangeError(this.name, 'Count', 1, 100);
 
         // Delete the original message so that more messages can be bulk deleted
         await message.delete()
